worker-service/requirements.txt
mysql-connector-python==8.3.0
redis==5.0.1

